1.2.x/core/src/test/java/org/apache/shiro/authc/credential/AllowAllCredentialsMatcherTest.java
/*
 * Licensed to the Apache Software Foundation (ASF) under one
 * or more contributor license agreements.  See the NOTICE file
 * distributed with this work for additional information
 * regarding copyright ownership.  The ASF licenses this file
 * to you under the Apache License, Version 2.0 (the
 * "License"); you may not use this file except in compliance
 * with the License.  You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing,
 * software distributed under the License is distributed on an
 * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
 * KIND, either express or implied.  See the License for the
 * specific language governing permissions and limitations
 * under the License.
 */
package org.apache.shiro.authc.credential;

import static org.junit.Assert.assertTrue;
import org.junit.Test;


/**
 * @since Jun 10, 2008 4:35:27 PM
 */
public class AllowAllCredentialsMatcherTest {

    @Test
    public void testBasic() {
        assertTrue(new AllowAllCredentialsMatcher().doCredentialsMatch(null, null));
    }

}
